feat:添加节点AI助手交互
