+++ packages/canvas/container/src/components/CanvasAction.vue
@@ -111,7 +111,6 @@
           class="ai-component"
           @confirm="handleAIConfirm"
           @cancel="handleAICancel"
-          @close="closeAIHelper"
           @refresh="handleAIRefresh"
         ></AIConfirmDialog>
       </div>
@@ -157,22 +156,17 @@ import {
   getCurrentElement,
   querySelectById
 } from '../container'
-import {
-  useLayout,
-  useMaterial,
-  useCanvas,
-  useMessage,
-  getMetaApi,
-  useHistory
-} from '@opentiny/tiny-engine-meta-register'
+import { useLayout, useMaterial, useCanvas, useMessage } from '@opentiny/tiny-engine-meta-register'
 import { Popover } from '@opentiny/vue'
 import shortCutPopover from './shortCutPopover.vue'
 import CanvasAIChat from './CanvasAIChat.vue'
 import AIConfirmDialog from './AIConfirmDialog.vue'
 import AILoadingDialog from './AILoadingDialog.vue'
-import { jsonrepair } from 'jsonrepair'
-import * as jsonpatch from 'fast-json-patch'
-import axios from 'axios'
+import { chat } from '../services/agentServices'
+import { utils } from '@opentiny/tiny-engine-utils'
+import useAIChat from '../composables/useAIChat'
+
+const { deepClone } = utils
 
 // 工具操作条高度
 const OPTION_BAR_HEIGHT = 24
@@ -238,25 +232,23 @@ export default {
   },
   emits: ['remove', 'selectSlot', 'setting'],
   setup(props) {
+    const { pageState: _pageState, getNode } = useCanvas()
+
     const {
-      getPageSchema,
-      pageState: _pageState,
-      getNodeAIStatus: _getNodeAIStatus,
+      getNodeAIStatus,
       openNodeAIChat,
       closeNodeAIHelper,
       startNodeAILoading,
       cancelNodeAILoading,
-      completeNodeAILoading,
-      shouldShowNodeAIHelper,
       shouldShowNodeAIChat,
       shouldShowNodeAILoading,
       shouldShowNodeAIConfirm,
-      setCurrentSchema,
-      findJsonPatchPath
-    } = useCanvas()
-
-    const { fixMethods, schemaAutoFix, getJsonObjectString, isValidFastJsonPatch, jsonPatchAutoFix } =
-      getMetaApi('engine.service.robot')
+      confirmNodeAIAction,
+      updateNodeAIStatus,
+      cancelNodeAIAction,
+      applyAIPatches,
+      buildAIChatRequest
+    } = useAIChat()
 
     const remove = () => {
       removeNodeById(getCurrent().schema?.id)
@@ -343,16 +335,6 @@ export default {
       return config?.configure?.isModal
     })
 
-    // AI助手显示状态
-    const shouldShowAIHelper = computed(() => {
-      const currentSchema = getCurrent().schema
-      if (!currentSchema?.id) {
-        return false // 默认不显示
-      }
-
-      return shouldShowNodeAIHelper(currentSchema.id)
-    })
-
     // 是否显示AI聊天界面
     const shouldShowAIChat = computed(() => {
       const currentSchema = getCurrent().schema
@@ -391,7 +373,7 @@ export default {
         return
       }
 
-      const currentStatus = _getNodeAIStatus(currentSchema.id)
+      const currentStatus = getNodeAIStatus(currentSchema.id)
 
       if (currentStatus && currentStatus.state !== 'hidden') {
         // 如果AI助手当前不是隐藏状态，则关闭它
@@ -762,69 +744,20 @@ export default {
       fixStyle.value = optionStyleValue
     })
 
-    const setSchema = (schema) => {
-      const { importSchema, setSaved } = useCanvas()
-      importSchema(schema)
-      setSaved(false)
-    }
-
     // AI聊天完成处理
-    const handleAIChatComplete = async (result) => {
+    const handleAIChatComplete = async (content) => {
       const currentSchema = getCurrent().schema
       if (!currentSchema?.id) {
         return
       }
+      const params = await buildAIChatRequest(content)
 
       // 先进入加载状态
       startNodeAILoading(currentSchema.id, 'AI正在处理您的请求...')
 
-      const response = await axios(result)
-      if (response.data) {
-        const content = response.data.choices[0].message.content
-        // let content = getJsonObjectString(streamContent)
-        // let jsonPatches = []
-        // try {
-        //   if (!isFinal) {
-        //     content = jsonrepair(content)
-        //   }
-        //   jsonPatches = JSON.parse(content)
-        // } catch (error) {
-        //   if (isFinal) {
-        //     logger.error('parse json patch error:', error)
-        //   }
-        //   return { isError: true, error }
-        // }
-
-        // // 过滤有效的json patch
-        // if (!isFinal && !isValidFastJsonPatch(jsonPatches)) {
-        //   return { isError: true, error: 'format error: not a valid json patch.' }
-        // }
-
-        const validJsonPatches = JSON.parse(content)
-        const parentPath = findJsonPatchPath(getPageSchema(), currentSchema.id)
-        const newSchema = validJsonPatches.reduce((acc, patch) => {
-          try {
-            // 创建新的 patch 对象，拼接完整路径
-            const fullPatch = {
-              ...patch,
-              path: parentPath + patch.path
-            }
-            return jsonpatch.applyPatch(acc, [fullPatch], false, false).newDocument
-          } catch (error) {
-            return acc
-          }
-        }, getPageSchema())
-        // schema纠错
-        fixMethods(newSchema.methods)
-        schemaAutoFix(newSchema.children)
-
-        // 更新Schema
-        setSchema(newSchema)
-        // if (isFinal) {
-        useHistory().addHistory()
-        // }
-        completeNodeAILoading(currentSchema.id)
-      }
+      const response = await chat(params)
+      // AI运行完：设置chatContent、aiModifiedNodeData，修改画布schema为AI的schema
+      applyAIPatches(currentSchema.id, response, content)
     }
 
     // AI加载取消处理
@@ -839,27 +772,51 @@ export default {
     }
 
     // 刷新AI操作（重新生成）
-    const handleAIRefresh = () => {
+    // 逻辑：修改画布节点schema为originalNodeData，设置aiModifiedNodeData为空，重新发起请求
+    const handleAIRefresh = async () => {
       const currentSchema = getCurrent().schema
       if (!currentSchema?.id) {
         return
       }
 
-      // 这里应该调用AI重新生成的逻辑
-      // 目前暂时先记录一个操作历史，然后关闭确认状态返回聊天状态
-      const { addNodeAIActionHistory } = useCanvas()
+      const nodeId = currentSchema.id
+      const currentAIStatus = getNodeAIStatus(nodeId)
+      if (!currentAIStatus) {
+        return
+      }
 
-      // 记录刷新操作
-      addNodeAIActionHistory(currentSchema.id, 'refresh', {
-        timestamp: Date.now(),
-        reason: '用户请求重新生成AI建议'
+      // 恢复画布节点schema为originalNodeData
+      if (currentAIStatus.originalNodeData) {
+        const node = getNode(nodeId)
+        if (node) {
+          Object.assign(node, deepClone(currentAIStatus.originalNodeData))
+          useMessage().publish({ topic: 'schemaChange', data: { nodeId } })
+        }
+      }
+
+      // 设置aiModifiedNodeData为空
+      updateNodeAIStatus(nodeId, {
+        aiModifiedNodeData: undefined
       })
 
-      // 返回到聊天状态，让用户重新输入
-      useCanvas().cancelNodeAIAction(currentSchema.id)
+      // 重新进入加载状态
+      startNodeAILoading(nodeId, 'AI正在重新生成...')
 
-      // 可以在这里触发一些回调，比如重新调用AI API
-      // 暂时先不实现具体的重新生成逻辑，只做状态切换
+      // 使用上次的聊天消息重新发起请求
+      const chatContent = currentAIStatus.chatContent
+      if (!chatContent) {
+        cancelNodeAILoading(nodeId)
+        return
+      }
+      const params = await buildAIChatRequest(chatContent)
+
+      try {
+        const response = await chat(params)
+        // AI运行完操作和 handleAIChatComplete 一样
+        applyAIPatches(nodeId, response, chatContent)
+      } catch (error) {
+        cancelNodeAILoading(nodeId)
+      }
     }
 
     // 确认AI操作
@@ -869,8 +826,7 @@ export default {
         return
       }
 
-      useCanvas().confirmNodeAIAction(currentSchema.id)
-      // 这里可以添加实际应用AI修改的逻辑
+      confirmNodeAIAction(currentSchema.id)
     }
 
     // 取消AI操作
@@ -880,7 +836,7 @@ export default {
         return
       }
 
-      useCanvas().cancelNodeAIAction(currentSchema.id)
+      cancelNodeAIAction(currentSchema.id)
     }
 
     return {
@@ -902,7 +858,6 @@ export default {
       labelStyle,
       labelRef,
       openAIHelper,
-      shouldShowAIHelper,
       shouldShowAIChat,
       shouldShowAILoading,
       shouldShowAIConfirm,
@@ -1209,12 +1164,24 @@ export default {
   .ai-component,
   .ai-component-loading {
     position: absolute;
-    bottom: -106px;
     right: 0;
     width: 360px;
+    opacity: 0;
+    transform: translateY(-10px);
+    animation: slideIn 0.2s ease-in-out forwards;
   }
   .ai-component-loading {
     width: 270px;
+  }
+}
+@keyframes slideIn {
+  from {
+    opacity: 0;
+    transform: translateY(-10px);
+  }
+  to {
+    opacity: 1;
+    transform: translateY(10px);
   }
 }
 </style>
